pyproject.toml
[project]
name = "simple-trade-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "oandapyv20>=0.7.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "six>=1.17.0",
]
